A lot of changes were made to make the Distance property be implementded correctly. Now we have an inital odometer of the vehicle inside the booking table that we add distance to, and then the returned distance updates the odometer inside the vehicles table

+++ CarRental.Common/Classes/Car.cs
@@ -7,6 +7,6 @@ public class Car : Vehicle
 {
 
 
-    public Car(int id, string regNo, string make, double odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus) :
+    public Car(int id, string regNo, string make, int odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus) :
        base (id, regNo, make, odometer, costKm, vehicleType, vehicleStatus) {}
 }
