+++ CarRental.Common/Classes/Motorcycle.cs
@@ -7,6 +7,6 @@ public class Motorcycle : Vehicle
 {
 
 
-    public Motorcycle(int id, string regNo, string make, double odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus) :
+    public Motorcycle(int id, string regNo, string make, int odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus) :
         base(id, regNo, make, odometer, costKm, vehicleType, vehicleStatus) {}
 }
